requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
Django==5.1.7
django-dynamic-fixture==4.0.1
django-filter==25.1
django_csp==3.8
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
